[UPD] movedout execution mode, init_loop & web_server
