+++ main/models/init_loop.h
@@ -1,0 +1,5 @@
+#pragma once
+
+/* The init-loop (run-once) execution model, defined in init_loop.c: run the script top-level, then
+ * drive setup()/loop() if defined. Used when PHP_PROJECT_WEB_SERVER is off. */
+void run_init_loop(const char *script);
